Delete menu_complete from .zshrc
This option will instantaniously select an option if
you press <Tab> instead of completing it just for the part
that was ambiguous

if there were options

- fooa
- foob
- fooc

with menu_complete, it will complete to `fooa` after one <Tab>.

without menu_complete, it will complete to `foo`, and only to `fooa`
after pressing <Tab> another time.

This is annoying if you have tons of folders named similiarly.